Add build CI pipeline setup
We're going to use our concourse build CI to tag each merge with a
version number. This is required to keep on consistency across all
environments. The tagging pipeline, also runs integration tests that are
setup on each repo. We're adding that configuration here, but making it
rather noop as we don't currently have integration tests and the unit
tests are run on travis.