Do not retrieve unnecesary envelopes
When querying the log-cache service, we only need to retrieve
the sourceID of tenant objects, not internal cloudfoundry services,
and we only want to get Counters and Gauges.

The tenants sourceID are always UUIDs, meanwhile the internal CF
services are strings for the different jobs. We parse the sourceID
and only download the ones that look like UUIDs.