Make server tests more blackbox
Some of our tests embed a reasonable level of implementation knowledge.
This change reduces that awareness a bit, by devolving the middleware
chain ordering to the server's `run` method that's actually used to
serve the app.

It's still slightly smelly, as it still requires the tests to be in the
same package as the implementation, so we can reach into the private
`server` struct and use its router. But that's not a /change/, so we end
up strictly better than before.

We also git+cf ignore the binary created when running tests.