Tell a metrics consumer if their auth creds might be invalid
We can't validate the credentials the user presents to us, hence we rely
on the upstream library and upstream service to do that for us.

To avoid leaking information, the combination of the upstream library
and service don't give us definitive information that a rejected request
failed because it had invalid auth creds. Instead, we have to extract
the response code and translate an upstream 404 into a message pointing
the metrics consumer in the right direction.

We can't translate this into a definitive HTTP 401 and message as we
don't know that this is the case. Passing on the upstream 404 with a
(hopefully) more helpful error message is the best we can achieve -
unless we want to validate the credentials before dispatching the
metrics requests, which seems like a runtime delay/load we wouldn't want
to impose on the metrics consumer or the platform.